move components base classes in components scope
