+++ stonks/__init__.py
@@ -1,6 +1,6 @@
-from stonks.app import App                                      # noqa
-from stonks.settings import Settings                            # noqa
-from stonks.components.fetchers.base_fetcher import Fetcher     # noqa
-from stonks.components.parsers.base_parser import Parser        # noqa
-from stonks.components.writers.base_writer import Writer        # noqa
-from stonks.components import manager                           # noqa
+from stonks.app import App                           # noqa
+from stonks.settings import Settings                 # noqa
+from stonks.components.base_fetcher import Fetcher   # noqa
+from stonks.components.base_parser import Parser     # noqa
+from stonks.components.base_writer import Writer     # noqa
+from stonks.components import manager                # noqa
