+++ stonks/components/__init__.py
@@ -2,3 +2,10 @@ from stonks.components import fetchers  # noqa
 from stonks.components import parsers   # noqa
 from stonks.components import writers   # noqa
 from stonks.components import manager   # noqa
+
+
+"""
+from stonks.components.base_fetcher import Fetcher
+from stonks.components.base_parser import Parser
+from stonks.components.base_writer import Writer
+"""
