+++ stonks/components/base_fetcher.py
