+++ stonks/components/base_parser.py
