+++ stonks/components/filename.py
